+++ angular_compiler/lib/angular_compiler.dart
@@ -1,6 +1,8 @@
 export 'src/analyzer.dart'
     show
         getTypeName,
+        linkToReference,
+        linkTypeOf,
         urlOf,
         ComponentReader,
         DependencyElement,